refactor: modularize render helpers
