+++ svg-time-series/src/chart/render/dimensions.ts
@@ -1,0 +1,20 @@
+import { BaseType, Selection } from "d3-selection";
+import { AR1Basis } from "../../math/affine.ts";
+
+export function createDimensions(
+  svg: Selection<BaseType, unknown, HTMLElement, unknown>,
+) {
+  const node: SVGSVGElement = svg.node() as SVGSVGElement;
+  const div: HTMLElement = node.parentNode as HTMLElement;
+
+  const width = div.clientWidth;
+  const height = div.clientHeight;
+
+  svg.attr("width", width);
+  svg.attr("height", height);
+
+  const bScreenXVisible = new AR1Basis(0, width);
+  const bScreenYVisible = new AR1Basis(height, 0);
+
+  return { width, height, bScreenXVisible, bScreenYVisible };
+}
